Fixed a bug in Sentry::force_login() that was storing the user id as a string, which caused a Sentry\SentryException to be thrown from Sentry::user().

+++ sentry.php
@@ -269,7 +269,7 @@ class Sentry
 
 		if ( ! is_int($id))
 		{
-			$id = static::user($id)->get('id');
+			$id = (int) static::user($id)->get('id');
 		}
 
 		Session::put(Config::get('sentry::sentry.session.user'), $id);
